refactor: split scans.py into modules
Extract the 2500+ line scans.py into focused modules:

- _scan_queries.py: low-level DB batch queries (findings, vulns, packages)
- _scan_diff.py: diff computation and list-view serialisation
- _scan_cache.py: scan_diff_cache read/write/invalidate/recompute
- _scan_helpers.py: shared helpers for scan trigger endpoints
- scan_triggers.py: Grype/NVD/OSV trigger and status routes

Update settings.py, packages.py, merger_ci.py, and scan_triggers.py
to import directly from the new modules instead of the monolithic
scans.py. Backward-compatible re-exports are kept for test imports.
Add tests covering _scan_cache error paths and _scan_diff edge cases.
